Support Matrix -> XMPP edits
