+++ changelog.d/154.feature
@@ -1,0 +1,1 @@
+Support Matrix -> XMPP edits
